+++ src/server/pokemon/index.ts
@@ -1,0 +1,3 @@
+"use server";
+
+export { getPokemonList as getPokemon, getPokemonById } from "./get";